Create hash_maps_flowers.py
The language of the flowers has a long history and has often been a topic resigned to the domain of dusty books in a thrift bookseller or a library.
With Flowerz, we want to give people lightning fast access to all of the possible meanings of their favorite flowers.
In this project, we are going to implement a hash map to relate the names of flowers to their meanings. 
In order to avoid collisions when our hashing function collides the names of two flowers, we are going to use separate chaining. 
We will implement the Linked List data structure for each of these separate chains.